==> 01-Intro.qmd <==
---
title: "Intro"
author: "DME"
format: html
editor: visual
---

## Coasta Wetland Monitoring Program (CWMP) - Laurentian Great Lakes Basin

To facilitate collaboration and information sharing between public, private, and government agencies throughout the Great Lakes basin, we developed standardized methods and indicators used for assessing wetland condition. Using an ecosystem approach and a stratified random site selection process, birds, anurans, fish, macroinvertebrates, vegetation, and physico-chemical conditions were sampled in coastal wetlands of all five Great Lakes including sites from the United States and Canada. The primary objective of this monitoring program was to implement a standardized standardized basin-wide coastal wetland monitoring program that would be a powerful tool to inform decision-makers on coastal wetland conservation and restoration priorities throughout the Great Lakes basin.

## Anuran Monitoring

The overall goal of the anuran monitoring program was to document the species present at the respective wetlands during the primary frog and toad mating seasons, which vary among species and from year to year depending on weather conditions between March and July. Given the massive area of the Great Lakes coastline, this monitoring program did not attempt to verify breeding productivity of frog and toad species within these wetlands, which would have required intensive searches for eggs or emergence of individuals from the wetland.

Anuran point counts within each wetland were completed three times:

1\) when nighttime air temperatures consistently reached 5C, usually in March or April;

2\) when nighttime temperatures consistently reached 10C and at least 15d from the first sample period (generally April or May); and

3\) when nighttime temperatures consistently reached 17C and at least 15d from the second sample period (generally June or July).

Anuran data were gathered using timed, unlimited-distance counts (full circle) at each point. Point surveys were completed from 0.5 h before to 4.5 h after sunset. Counts were not conducted when winds were high (\> 20 km/h) or during rain, although sampling during periods of drizzle or light winds was acceptable, especially if anuran calling was deemed normal.

Each anuran survey was 3 min in duration. Data recorded at each survey point before or immediately after the survey included the following:

1\) point ID with a recording of the GPS waypoint,

2\) date,

3\) start time,

4\) name of observers (for safety reasons we recommend two individuals participate in the counts), and

5\) weather conditions including air temperature, wind speed and direction, precipitation, cloud cover, ambient noise levels, and if possible water temperature.

Data recorded during the survey included:

1\) identified species;

2\) calling intensity (see below) of all frog and toad species,

3\) the nearest distance of each anuran detection in one of three categories (\<50 m, 50-100 m, or \>100 m), and

4\) location of each detection with reference to the observer (semicircle in front of or behind the observer, oriented at a fixed direction facing the wetland).

For safety reasons, we recommend only shore-based and no over-water nighttime surveys. Calling intensity for each species was coded according to three categories: Code 1 described a condition where calls were not simultaneous and individuals of a given species could be counted; Code 2 represented a level of activity where some calling was simultaneous, but numbers of individuals could be reliably estimated; Code 3 described a full chorus with so many continuous and overlapping calls that individuals could not be accurately counted.

## Resources

For more details on sampling, please review Uzerski et al. 2017. Standard measures of Coastal Wetland Condition: Implementation at a Laurentian Great Lakes Basin-Wide Scale. Wetlands. 37: 15-32.

A copy of this data collection protocol and meta data form can be found in the `Resource` folder in this project directory.

## Author

The original R code were created by Danielle Ethier, Population Scientist with Birds Canada. Modification Date: December 2023

Great Lakes Coastal Wetland Monitoring Program (GLCWMP), funded by the Great Lakes Restoration Initiative through the US EPA Great Lakes National Program Office to D. Uzarski, Central Michigan University. 2011-2025. Grant numbers GL-00E00612-0, 00E01567 and 00E02956. URL: https://www.greatlakeswetlands.org
